Move Korean shell components out of src/pages/ to fix the build
The previous commit's rename to src/pages/ko/_components/ did not actually skip the subtree. Gatsby's page validator only checks the file basename, not parent directories:

  packages/gatsby-page-utils/src/validate-path.ts (v4.25.9)
  return (
    parsedPath.name.slice(0, 1) !== `_` &&
    ...
  )

So src/pages/ko/_components/KoreanDesktop/lottieAnimations.ts still got picked up as a page component and the build kept failing with the same error.

The Gatsby-convention fix: helper components belong under src/components/, not src/pages/. Move the Korean shell components to src/components/Korean/ and update import paths.

Inner relative imports change because the depth shifted:
- '../../../../context/{App,Toast,Window}' -> '../../../context/{App,Toast,Window}'
- '../../../../navs' -> '../../../navs'
- '../../_translations' -> '../../../pages/ko/_translations'

Outer imports updated in gatsby-browser.tsx, gatsby-ssr.js, and src/pages/ko/_KoreanHome.tsx.

The /ko page still works the same way at runtime; only the file layout changes.

Generated-By: PostHog Code
Task-Id: e481a749-e4c5-44d5-8bec-4ec278e15255
