+++ src/components/Korean/KoreanDesktop/DraggableDesktopIcon.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useRef } from 'react'
 import { motion, useDragControls } from 'framer-motion'
 import { AppLink, AppItem } from 'components/OSIcons/AppIcon'
 import ZoomHover from 'components/ZoomHover'
-import { useApp } from '../../../../context/App'
+import { useApp } from '../../../context/App'
 
 interface DraggableDesktopIconProps {
     app: AppItem
