+++ src/components/Korean/KoreanDesktop/HourglassSpinner.tsx
