+++ src/components/Korean/KoreanDesktop/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import Link from 'components/Link'
-import { useApp } from '../../../../context/App'
+import { useApp } from '../../../context/App'
 import { IconDemoThumb, AppIcon } from 'components/OSIcons'
 import { AppItem } from 'components/OSIcons/AppIcon'
 import ContextMenu from 'components/RadixUI/ContextMenu'
@@ -13,9 +13,9 @@ import useTheme from 'hooks/useTheme'
 import { motion } from 'framer-motion'
 import HedgeHogModeEmbed from 'components/HedgehogMode'
 import ReactConfetti from 'react-confetti'
-import { useToast } from '../../../../context/Toast'
+import { useToast } from '../../../context/Toast'
 import usePostHog from 'hooks/usePostHog'
-import { translateKo } from '../../_translations'
+import { translateKo } from '../../../pages/ko/_translations'
 import KoreanCookieBannerToast from '../KoreanCookieBannerToast'
 
 declare global {
