+++ src/components/Korean/KoreanDesktop/lottieAnimations.ts
