+++ src/components/Korean/KoreanHeroCarousel/slides.tsx
