+++ src/components/Korean/KoreanMDXViewer.tsx
